+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- App logos beside them in `CustomizeView`
+
 ## [0.0.3] - 2020-07-28
 
 ### Added