+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.0.5] - 2020-08-10
 
 ### Added
 
@@ -89,7 +89,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Windows Support (Script Part Fully Done)
 - Fonts: (Roboto, Roboto Mono)
 
-[unreleased]: https://github.com/YazeedAlKhalaf/Split_It/compare/v0.0.4...HEAD
+[unreleased]: https://github.com/YazeedAlKhalaf/Split_It/compare/v0.0.5...HEAD
+[0.0.5]: https://github.com/YazeedAlKhalaf/Split_It/releases/tag/v0.0.5
 [0.0.4]: https://github.com/YazeedAlKhalaf/Split_It/releases/tag/v0.0.4
 [0.0.3]: https://github.com/YazeedAlKhalaf/Split_It/releases/tag/v0.0.3
 [0.0.2]: https://github.com/YazeedAlKhalaf/Split_It/releases/tag/v0.0.2